refactor: split plan detail into 3 components
