+++ src/routes.tsx
@@ -18,7 +18,7 @@ import MyAccount from './components/myAccount/'
 import PaymentDetail from './components/payment/detail'
 import PaymentList from './components/payment/list'
 import PricePlanList from './components/plan'
-import PlanDetail from './components/plan/detail'
+import PlanDetail from './components/plan/detail/index'
 import { ReportPage } from './components/report'
 import Settings from './components/settings/index'
 import WebhookLogs from './components/settings/webHooks/webhookLogs'
