+++ tsconfig.json
@@ -4,6 +4,7 @@
     "useDefineForClassFields": true,
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "module": "ESNext",
+    "forceConsistentCasingInFileNames": true,
     "skipLibCheck": true,
     "rootDir": "./src",
     "outDir": "./dist",
